Rename Unity example MexGrpcSample to MexRestSample
