+++ edge-mvp/unity/examples/MexRestSample/Assets/Logos.meta
