+++ edge-mvp/unity/examples/MexRestSample/Assets/Plugins.meta
