+++ edge-mvp/unity/examples/MexRestSample/Assets/Plugins/MatchingEngineSDKRestLibrary.dll.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cc3d9de211e5c434080d69539cf8a64f
+guid: 9bbd2d7a474464f8e9fa5fa3814043f3
 PluginImporter:
   externalObjects: {}
   serializedVersion: 2
