+++ edge-mvp/unity/examples/MexRestSample/Assets/Scenes/MexScene.unity.meta
