+++ edge-mvp/unity/examples/MexRestSample/Assets/StatusContainer.cs
