+++ edge-mvp/unity/examples/MexRestSample/Packages/manifest.json
