+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/DynamicsManager.asset
