+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/EditorBuildSettings.asset
@@ -4,5 +4,8 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes: []
+  m_Scenes:
+  - enabled: 1
+    path: Assets/Scenes/MexScene.unity
+    guid: b00e986c5233b4b699a1961c591f934d
   m_configObjects: {}
