+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/GraphicsSettings.asset
