+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/NavMeshAreas.asset
