+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/ProjectSettings.asset
