+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/ProjectVersion.txt
