+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/QualitySettings.asset
