+++ edge-mvp/unity/examples/MexRestSample/ProjectSettings/VFXManager.asset
